Issue #17 fix
Remove `inDictionary:` from the class creation templates. Add 2 new tests:
  RwHybridBrowserToolTest>>testHybridClassCreationWithClassCreationTemplateA
  RwHybridBrowserToolTest>>testHybridClassCreationWithClassCreationTemplateB

to fill the gap in test coverage ... there were a few other tweaks needed for the templates to be correct ... mainly in the area of using hybrid naming when it's not appropriate and expand the options for creating templaces of variable and bytes classes.

tests are green:

  87 run, 86 passes, 1 expected defects, 0 failures, 0 errors, 0 unexpected passes