Fix for Jadeite issue #507 - hit deprecated method.
Jadeite sends ``canUnderstand: #comment`` method. If ``comment`` is not available, try to get the ``description``.  In this case, I think the problem was that a meta class was being refreshed and the meta class didn't understand ``comment`` even though the comment method was available.

I believe the fix is to make sure to use the non-meta class when sending ``canUnderstand:``.