Fix Jadeite issue 717 inspector mnu
The windows registry holds a dictionary of client window handles to root objects of those windows - the windows model.

An association gets added to the dictionary when the window is created but is also used for saving the result object of ad-hoc operations within those windows. The root object was getting replaced during the course of the ad-hoc operation thus freeing it to be garbage collected. Now, the value of the association is an array and we we add the result object of the ad-hoc operation to that array.