Fixes Jadeite issue #832, find method containing fail
This appears to be a change in 3.6. From memory, `ClassOrganizer >> substringSearch:` used to return two arrays - the methods & the first offset of the search string. Now three arrays are returned. `RowanQueryService>>methodsContaining:` accessed the "last" element of the array to get the first offset. It should have referenced the second element.