Issue #783:
If category and packageName properties in Trait are equal only save category to disk.
Completely eliminate the code involved with the shebang property.
Rename #_owner message for GsNMethod (accessing the DEBUGINFO_packageInfo slot in
  GsNMethod debugInfo) which will eventually replace usage of #_rowanPackageInfo message
  since we will be using that field for storing either Trait information or loaded
  method information for the GsNMethod (they are not expected to coexist).
Add #isRowanLoadedMethod to be used along with #isTrait to determine what is stored in
  the _owner field in compiledMethod (formerly called _rowanPackageInfo).
All of the places in Rowan 3 that used to call #_rowanPackageInfo
  switched to send #_owner to make it easier to eliminate
  _rowanPackageInfo completely (when Rowan 2 changed or no longer
  supported)